services/cortex-sentinel/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-sentinel"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "nats-py==2.7.0",
    "psutil==5.9.8",
    "httpx==0.27.0",
    "pydantic==2.7.1",
    "hvac==2.1.0",
    "cryptography==42.0.5",
    "structlog==24.1.0",
    "prometheus-client==0.20.0",
    "opentelemetry-api==1.24.0",
    "opentelemetry-sdk==1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.24.0",
]

[tool.uv]
dev-dependencies = [
    "pytest==8.1.1",
    "pytest-asyncio==0.23.6",
    "pytest-mock==3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["sentinel*"]
